scripts/requirements.txt
pdfplumber==0.11.4
python-docx==1.1.2
PyMuPDF==1.24.0
Pillow==10.4.0
